Refactor add/update fiche intervention
