+++ back/prisma/schema.prisma
@@ -882,7 +882,7 @@ model Bsff {
 }
 
 model BsffFicheIntervention {
-  id         String   @id
+  id         String   @id @default(cuid())
   createdAt  DateTime @default(now())
   updatedAt  DateTime @updatedAt
   numero     String
