+++ back/src/bsffs/database.ts
@@ -2,7 +2,6 @@ import type { SetRequired } from "type-fest";
 import { Bsff, BsffFicheIntervention, Prisma } from ".prisma/client";
 import { UserInputError } from "apollo-server-express";
 import prisma from "../prisma";
-import { getFicheInterventionId } from "./converter";
 
 export async function getBsffOrNotFound(
   where: SetRequired<Prisma.BsffWhereInput, "id">
@@ -21,21 +20,14 @@ export async function getBsffOrNotFound(
 }
 
 export async function getFicheInterventionBsffOrNotFound(
-  bsffId: string,
-  ficheInterventionNumero: string
+  where: SetRequired<Prisma.BsffFicheInterventionWhereInput, "id">
 ): Promise<BsffFicheIntervention> {
-  const ficheInterventionId = getFicheInterventionId(
-    bsffId,
-    ficheInterventionNumero
-  );
-  const ficheIntervention = await prisma.bsffFicheIntervention.findUnique({
-    where: {
-      id: ficheInterventionId
-    }
+  const ficheIntervention = await prisma.bsffFicheIntervention.findFirst({
+    where
   });
   if (ficheIntervention == null) {
     throw new UserInputError(
-      `La fiche d'intervention n°${ficheInterventionNumero} n'existe pas pour le bordereau n°${bsffId}.`
+      `La fiche d'intervention n°${where.id} n'existe pas.`
     );
   }
   return ficheIntervention;
