+++ back/src/bsffs/resolvers/mutations/__tests__/signBsff.integration.ts
@@ -15,7 +15,6 @@ import {
   createBsffAfterEmission,
   createBsffBeforeTransport,
   createBsffBeforeReception,
-  createBsffAfterReception,
   createBsffAfterTransport,
   createBsffBeforeOperation
 } from "../../../__tests__/factories";
