+++ back/src/bsffs/resolvers/mutations/createFicheInterventionBsff.ts
@@ -1,0 +1,30 @@
+import { checkIsAuthenticated } from "../../../common/permissions";
+import { MutationResolvers } from "../../../generated/graphql/types";
+import prisma from "../../../prisma";
+import {
+  flattenFicheInterventionBsffInput,
+  unflattenFicheInterventionBsff
+} from "../../converter";
+import { ficheInterventionSchema } from "../../validation";
+
+const addFicheInterventionBsff: MutationResolvers["addFicheInterventionBsff"] = async (
+  _,
+  { input },
+  context
+) => {
+  checkIsAuthenticated(context);
+
+  const ficheInterventionData = flattenFicheInterventionBsffInput(input);
+
+  await ficheInterventionSchema.validate(ficheInterventionData, {
+    abortEarly: false
+  });
+
+  const ficheIntervention = await prisma.bsffFicheIntervention.create({
+    data: ficheInterventionData
+  });
+
+  return unflattenFicheInterventionBsff(ficheIntervention);
+};
+
+export default addFicheInterventionBsff;
