+++ back/src/bsffs/resolvers/mutations/updateFicheInterventionBsff.ts
@@ -1,4 +1,3 @@
-import { UserInputError } from "apollo-server-express";
 import { checkIsAuthenticated } from "../../../common/permissions";
 import { MutationResolvers } from "../../../generated/graphql/types";
 import prisma from "../../../prisma";
@@ -7,31 +6,18 @@ import {
   unflattenFicheInterventionBsff
 } from "../../converter";
 import { ficheInterventionSchema } from "../../validation";
-import {
-  getBsffOrNotFound,
-  getFicheInterventionBsffOrNotFound
-} from "../../database";
-import { isBsffContributor } from "../../permissions";
+import { getFicheInterventionBsffOrNotFound } from "../../database";
 
 const updateFicheInterventionBsff: MutationResolvers["updateFicheInterventionBsff"] = async (
   _,
-  { id, numero, input },
+  { id, input },
   context
 ) => {
-  const user = checkIsAuthenticated(context);
-  const bsff = await getBsffOrNotFound({ id });
-  await isBsffContributor(user, bsff);
+  checkIsAuthenticated(context);
 
-  if (bsff.emitterEmissionSignatureDate) {
-    throw new UserInputError(
-      `Il n'est pas possible d'éditer une fiche d'intervention après la signature de l'émetteur`
-    );
-  }
-
-  const existingFicheIntervention = await getFicheInterventionBsffOrNotFound(
-    id,
-    numero
-  );
+  const existingFicheIntervention = await getFicheInterventionBsffOrNotFound({
+    id
+  });
   const ficheInterventionData = flattenFicheInterventionBsffInput(input);
 
   await ficheInterventionSchema.validate({
